feat(chat): add `include` role to include files
Files may be included in the chat by a special `include` role. Each
file's contents will be added to an additional `user` role message with
the files separated by `==> {path} <==` where `{path}` is the path to
the file. Globbing is expanded out via `glob.glob` and relative apths to
the current working directory (as determined by `getcwd()`) will be
resolved to absolute paths.

Example:

```
>>> user

Generate documentation for the following files

>>> include

/home/user/myproject/src/../requirements.txt
/home/user/myproject/**/*.py

```

Fixes: #69